Make the midi play command an ApplicationCommand
This tweaks the playback command to be an application command, so that
there's no potential for triggering multiple simultaneous playbacks
at once (for now).

This also uncouples the command from requiring that the current file be
a midi file, as it can now accept a filename argument which will play
instead of the current file.

Playback now also stops and errors are caught if files are broken.