Add ability to change midi program
Allow the user to select a midi program to be used for playback. The
selection is persisted and re-applied at plugin load time and when the
port is reset.

The programs are stored in a JSON file and displayed to the user in
the same banks as the general MIDI program set, so those familiar should
be able to find what they want easily.