Refactor settings into a function call
This replaces the global piano_prefs settings object with a similarly
named function which takes an argument that is the setting that you'd
like to get. Default values are stored in the plugin code so that there
is always a sensible value even if the preferences file is missing.

The new function can take an optional second argument, in which case the
setting value is updated.