Include a new piano_layout setting
This new setting will ultimately control what layout is used by default
for the piano, since the ultimate plan is to have more than one layout
available. 

The setting is implemented here, but since some glue code is still 
to be written, it's only used in one place currently and the other
location that needs to know the layout has a hard coded value (it will
be replaced with a custom command).