create new get_piano_view() method
This method can find, create and focus the piano view, as well as change
the piano layout that it's using.

This works under the assumption that regardless of the number of windows
there should be exactly one piano view available. This method can create
it if it's missing, optionally also focus it for display (even raising
the window as needed) and can swap the existing layout for a new one
on the fly.

The piano view is now forced to be read-only and also carries a setting
that indicates that it is the piano, which allows the key binding for
playing via keyboard to trigger regardless of location.

Tangentially there is also a new helper for getting package resource
names in a package-name friendly way, which is used in various places.