Implement a new show_piano command
This implements a new command that is a thin wrapper around the new 
get_piano_view() method. It allows displaying and focusing the piano
view, and can also dynamically swap the piano layout as well.

As currently defined it would be possible to have multiple command
palette entries or even key bindings that can focus or swap the layout
as needed.