Change to new layout extension for piano layouts
This finalizes the initial intentions for having multiple layouts for
the piano by assigning the layouts their own extension and modifying the
syntax definition to suit.

The code for loading up the layout is modified to suit, and a second
layout (the 6 octave piano used for testing) is also included.