fix bug when the last token in the selection is a note
when playing `piano-tune`s or toggling notation, if the last token was a note, then it would try to get the next token to check if it is a sharp, but there are no more tokens, so it crashed. This works around that by checking if the token is None.