Include a new Display Driver class
This encapsulates some of the functionality from the Piano class and
abstracts it away into a class that can drive a piano display.

Instead of every note request updating the display immediately, instead
the requests go into a queue and trigger an out-of-band update to happen
at a point in the future. The update is scheduled when the first request
is made, and all requests that occur while the timeout is happening are
added to the eventual update without triggering a new one.

Once the timeout hits, the display is updated wholesale, modifying the
state of keys as appropriate. If the state of a key ends up being in the
same state as it was in prior to any updates, then that key is left in
that state.

This ensures that on fast passages, at the end of the timeout there's
only a single update that brings the key in line with what the final
state should be.