Update the piano event listener
This event listener attaches directly to the piano view, so it can
create a display driver when an instance gets created that updates the
display using the new mechanism.

This also covers playback of midi input coming from an external device
and midi file playback, both of which go through asking the piano to
update itself.