Update the piano tune event listener
This directs display requests to the piano event listener (if any) in
order to update the piano instead of using it's methods to update the
display. The request doesn't ask the note to play so that it's possible
to play a tune without the piano being displayed.

With this change, the methods that were refactored into the new display
driver class can now be removed.