Include a new setting to control update frequency
Instead of hard coding a display frequency of 30fps, allow the user to
control the number of updates that occur per second; this allows for
tuning based on what you're playing back, CPU specs, etc.

This is currently constrained between 1-1000 inclusive; setting this too
low and trying to play back manually is probably not a great idea.

The default value of 20 updates per second seems to be a good sweet spot
for snappy updates without overwhelming the CPU.