Modify play_midi_file to allow data URI's
This slight modification to the play_midi_file command allows it to
detect when the filename that you pass it is a data URI, which will
be decoded and played back.

This allows you to inline playback directly into keybindings and such
without having to have the actual file laying around.