experiments with recording notes played on the piano into piano-tune
for recording to work, there must be exactly 2 ST layout groups open in the window, and one of those active tabs should be the piano and one should be a piano-tune whose first line starts with `// recording` or `// recording with octaves`

then, all notes played will be output into the piano-tune view. Note that it doesn't support simultaneous notes yet, and probably can't due to where it is currently implemented.
Nor does it record the note lengths atm.

If the comment contains `with octaves`, then notes from adjacent octaves will be marked in the piano-tune file with the usual `>` or `<` tokens, and if the octave is further away from the last note in the file/view before the caret, it will record an absolute octave instruction.

The idea is that you can quickly open up a piano-tune tab, add the comment and play on the piano (or play a midi file - currently untested, and probably has abysmal performance) to record your song (or help convert from midi to piano-tune format).