Fix line ending issues when showing the piano
Since we're using load_resource() and `append` to load and display the
piano, we run into line ending issues if the package's resource isn't
using the same line endings as the user has set for their platform.

Here we normalize the input to use only `\n` the same way that the
`exec` command does, which resolves that issue.